The navbar will now show the correct tab as active
The navbar should now show the correct tab as active.
I did this by making the navbar template set the class of its
 list items equal to clearly named variables that should be
either empty or contain the string 'active'.
The template should then be rendered with the correct
variable holding a value of 'class' and the correct
table will be rendered as active.